Fill out API, add docs, split tests.
