+++ .gitignore
@@ -1,5 +1,6 @@
 lib/
 coverage/
+docs/
 
 # Created by https://www.gitignore.io/api/osx,node,grunt,vim,linux
