+++ src/index.ts
@@ -1,1 +1,1 @@
-export {default as Item} from './Item'
+export * from './Item'
